+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.11.5] - 2025-03-01
 
 ### Changed
-- Changed location watching logic to always watch
+- Changed location watching logic to periodically request location
+- Increased filter menu gaps
 
 ## [0.11.4] - 2025-03-01